requirements.txt
discord.py==2.3.2
requests
beautifulsoup4
selenium
undetected-chromedriver
python-dotenv
